requirements.txt
# AI-Core / Frank — Python Dependencies
# Install: pip install -r requirements.txt

# --- Core Web Framework ---
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.0

# --- HTTP Clients ---
requests>=2.31
httpx>=0.25
aiohttp>=3.9

# --- Image / Vision ---
Pillow>=10.0
pytesseract>=0.3.10

# --- Document Processing ---
pypdf>=3.0
python-docx>=1.0

# --- Data / ML ---
numpy>=1.24
PyYAML>=6.0

# --- Security ---
cryptography>=41.0

# --- GUI (Tkinter overlay — ships with Python, no pip needed) ---
# tkinter is part of the Python standard library

# --- Optional: GTK tray icon & Writer app ---
# PyGObject>=3.42    # requires system libgirepository: apt install libgirepository1.0-dev
# PySide6>=6.5       # only needed for the Qt live-wallpaper module
